Merge two data.frames, but where there are common fields those in x are retained and those in y are dropped.

Merge two data.frames, but where there are common fields those in x are retained and those in y are dropped.

.merge_with_precedence(x, y, by, allow.cartesian = T, as_data_frame = T)

Arguments

x

the data.frame whose columns take precedence

y

another data.frame

by

a vector of column names to merge on

allow.cartesian

boolean indicating whether it's ok for repeated values in either table to merge with each other over and over again.

as_data_frame

boolean indicating whether to ensure the returned object is a data.frame instead of a data.table. This ensures compatibility with GCT object conventions, that is, the rdesc and cdesc slots must be strictly data.frame objects.

Value

a data.frame or data.table object

See also