Write a nested list to a GMT file

Write a nested list to a GMT file

.write.gmt(lst, fname)

Arguments

lst
the nested list to write. See details.

fname
the desired file name

Details

lst needs to be a nested list where each sub-list is itself a list with the following fields: - head: the name of the data - desc: description of the corresponding data - len: the number of data items - entry: a vector of the data items

See also