Transform a GCT object in to a long form data.table (aka 'melt')

Utilizes the data.table::melt function to transform the matrix into long form. Optionally can include the row and column annotations in the transformed data.table.

melt.gct(
  g,
  suffixes = NULL,
  remove_symmetries = F,
  keep_rdesc = T,
  keep_cdesc = T,
  ...
)

Arguments

g
the GCT object

suffixes
the character suffixes to be applied if there are collisions between the names of the row and column descriptors

remove_symmetries
boolean indicating whether to remove the lower triangle of the matrix (only applies if g@mat is symmetric)

keep_rdesc
boolean indicating whether to keep the row descriptors in the final result

keep_cdesc
boolean indicating whether to keep the column descriptors in the final result

...
further arguments passed along to data.table::merge

Value

a data.table object with the row and column ids and the matrix values and (optinally) the row and column descriptors

See also