Reads a dataframe of Peptides IDs and their Scores and run KEA3

reads a dataframe of Peptides IDs and their Scores (LFC, p-value, ... etc) and run KEA3 on a subset of these peptides or all of them

read_kea(
  df,
  filter = T,
  cutoff = 0.2,
  cutoff_abs = T,
  direction = "higher",
  rm_duplicates = T,
  method = "MeanRank",
  lib = c("kinase-substrate"),
  ...
)

Arguments

df

dataframe, must have at least Peptide and Score columns

filter

boolean to subset peptides or not

cutoff

numeric to act as the cutoff to filter out peptides

cutoff_abs

boolean (use absolute value or not) default is TRUE

direction

("lower", "higher) filter based on lower than or higher than the cutoff values (default to "higher")

rm_duplicates

boolean (TRUE or FALSE) remove genes duplicates

method

"MeanRank" takes the mean rank across all libraries or "MeanFDR" takes the mean of FDR across all libraries (default is "MeanRank")

lib

searched kea libraries "kinase-substrate" or "all" (default is "kinase-substrate" which will return only kinase libraries like ChengKSIN, PTMsigDB, PhosDAll)

...,

arguments passed to rank_kinases function

Value

dataframe, Ranked and quartiled table

Details

This function a dataframe of Peptides IDs and their Scores (LFC, p-value, ... etc) and run KEA3 on a subset of these peptides or all of them