Write a GCT object to disk in GCT format

Write a GCT object to disk in GCT format

.write.gct(ds, ofile, precision = 4, appenddim = T, ver = 3)

Arguments

ds

the GCT object

ofile

the desired output filename

precision

the numeric precision at which to save the matrix. See details.

appenddim

boolean indicating whether to append matrix dimensions to filename

ver

the GCT version to write. See details.

Details

Since GCT is text format, the higher precision you choose, the larger the file size. ver is assumed to be 3, aka GCT version 1.3, which supports embedded row and column metadata in the GCT file. Any other value passed to ver will result in a GCT version 1.2 file which contains only the matrix data and no annotations.

See also